refactor: move lifecycle functions into separate file and readd shutdown functionality
